App/backend/requirements.txt
openai
pypdf
python-dotenv
numpy
tiktoken
fastapi
uvicorn[standard]
python-multipart




